Initial commit
This is a script that uses op, fzf, and wl-copy to quickly search for
and find 1password logins and copy those to a clipboard.

This could use a lot more work, mostly in making it more flexible in the
tools it uses, but it does what I need it to do.